moved MapContainer components inside components folder and created registration component
