+++ src/components/Registration.tsx
@@ -1,0 +1,13 @@
+import React, { FC, ReactElement } from 'react';
+
+
+const Registration: FC = () => {
+return (
+  <>
+  
+  </>
+)
+
+}
+
+export default Registration;
